Windows tree view install .md
Currently Windows users don't have a good option for Tree view. There are two options they have, they can use the default tree cmd that comes from windows, but it is heavy and wonky, and the displayed format is pretty ugly. The other option is to install a program that will display a nicely formatted tree view, but it takes quite a few extra steps. I have included instructions for both.