Fix for the 04 update terminal regarding wr.
Fixed the bug in the code for updating the terminal. Previously there was only an alias set, and that used the single quotes, but when I changed it to store the reference to the Windows Root (wr) in a variable and then use that variable in the alias, I forgot to take out the single quotes. This doc works, and have been verified in the video playlist I added to the first doc in a previous commit.