Clarified caveat 1 as it was incorrect.
"WSL does not have access to a GUI"

Technically there are ways to set this up, but for this guide and it's workflow, we assume we don't have access. Once I find a consistent, bug free, recommended, and easy to setup way to fix this I will update the guide, but ATM it seems too messy for such a small payoff, especially when we don't need access to a GUI at all anyway.